rename planets to planet
