+++ app/routes/planet_routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
-from ..models.planets import planets
+from ..models.planet import planets
 
 planets_bp = Blueprint("planets_bp",__name__,url_prefix="/planets")
 
@@ -15,6 +15,3 @@ def get_all_planets():
             galaxy=planet.galaxy
         ))
     return result_list
-
-    
-
